requirements.txt
# Web Framework
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.3

# Database
SQLAlchemy==2.0.39

# Machine Learning & Data Science
scikit-learn==1.5.0
numpy==1.26.4

# AI/LLM Integration
openai==1.69.0
litellm==1.55.8
sentence-transformers==5.0.0
faiss-cpu==1.11.0.post1

# HTTP Client
requests==2.32.3

# Image Processing & QR Code
Pillow==11.1.0
pyzbar==0.1.9

# Environment Variables
python-dotenv==1.1.1
